Update versions of peer dependencies to allow latest Vite and Vite Svelte plugin
Now that [Vite 5 has been released](https://vitejs.dev/blog/announcing-vite5), it would be nice to have te versions of the peer dependencies for it and the Svelte Vite plugin which uses Vite, updated to the latest versions. If this could be published as a new release that would most awesome!